Moved event objects to a sub-namespace inside Core\EventDispatcher
