+++ module/Core/src/EventDispatcher/Event/VisitLocated.php
@@ -2,26 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\EventDispatcher;
+namespace Shlinkio\Shlink\Core\EventDispatcher\Event;
 
-use JsonSerializable;
-
-final class VisitLocated implements JsonSerializable
+final class VisitLocated extends AbstractVisitEvent
 {
-    private string $visitId;
-
-    public function __construct(string $visitId)
-    {
-        $this->visitId = $visitId;
-    }
-
-    public function visitId(): string
-    {
-        return $this->visitId;
-    }
-
-    public function jsonSerialize(): array
-    {
-        return ['visitId' => $this->visitId];
-    }
 }
